added filters

+++ resource/database-resource/representation/documents.go
@@ -2,5 +2,5 @@ package representation
 
 type Documents struct {
 	Objects []interface{} `json:"objects"`
-	Total   int64         `json:"total"`
+	Total   int           `json:"total"`
 }
