+++ resource/database-resource/transformer/documents-to-view.go
@@ -2,7 +2,7 @@ package transformer
 
 import "github.com/MongoDBNavigator/go-backend/resource/database-resource/representation"
 
-func DocumentsToView(data []interface{}, total int64) *representation.Documents {
+func DocumentsToView(data []interface{}, total int) *representation.Documents {
 	result := new(representation.Documents)
 	result.Total = total
 	result.Objects = make([]interface{}, len(data))
